+++ fwdpp/extensions/regions.hpp
@@ -3,6 +3,8 @@
 #define __FWDPP_EXTENSIONS_REGIONS_HPP__
 
 #include <limits>
+#include <cassert>
+#include <stdexcept>
 #include <algorithm>
 #include <gsl/gsl_randist.h>
 #include <fwdpp/type_traits.hpp>
@@ -49,6 +51,18 @@ namespace KTfwd
 							    sbeg(std::move(__sbeg)),send(std::move(__send)),
 							    shmodels(std::move(__shmodels))
       {
+	if(nbeg.size()!=nend.size() || nbeg.size()!=nweights.size())
+	  {
+	    throw std::runtime_error("input vectors must all be the same size");
+	  }
+	if(sbeg.size()!=send.size() || sbeg.size()!=sweights.size())
+	  {
+	    throw std::runtime_error("input vectors must all be the same size");
+	  }
+	if(!sweights.empty() && sweights.size() != shmodels.size())
+	  {
+	    throw std::runtime_error("incorrect number of shmodels");
+	  }
 	if(nweights.size())
 	  nlookup = KTfwd::fwdpp_internal::gsl_ran_discrete_t_ptr(gsl_ran_discrete_preproc(nweights.size(),&nweights[0]));
 	if(sweights.size())
@@ -83,6 +97,11 @@ namespace KTfwd
 	\param recycling_bin A recycling bin for mutations
 	\param mutations A container of mutations
 	\param lookup Lookup table for mutations
+
+	\precondition If nmu == 0, then nbeg/nend cannot be empty.  Similarly,
+	if smu == 0, then sbeg,end cannot be empty.  These conditions are checked in debug 
+	mode via the assert macro.  It is up to the calling environment to prevent this situation
+	from arising.
       */
       template<typename queue_t,
 	       typename lookup_table_t,
@@ -98,11 +117,15 @@ namespace KTfwd
 	bool is_neutral = (gsl_rng_uniform(r) < nmu/(nmu+smu)) ? true : false;
 	if( is_neutral )
 	  {
+	    assert(!nbeg.empty());
+	    assert(!nend.empty());
 	    size_t region = gsl_ran_discrete(r,nlookup.get());
 	    double pos = posmaker(r,nbeg[region],nend[region],lookup);
 	    return fwdpp_internal::recycle_mutation_helper(recycling_bin,mutations,
 							   pos,0.,0.,generation);
 	  }
+	assert(!sbeg.empty());
+	assert(!send.empty());
 	size_t region = gsl_ran_discrete(r,slookup.get());
 	double pos = posmaker(r,sbeg[region],send[region],lookup);
 	return fwdpp_internal::recycle_mutation_helper(recycling_bin,mutations,
@@ -162,11 +185,19 @@ namespace KTfwd
 			  const std::vector<double> & __end,
 			  const std::vector<double> & __weight ) : beg(__beg),end(__end)
       {
+	if(beg.size() != end.size() || beg.size() != __weight.size())
+	  {
+	    throw std::runtime_error("input vectors must all be the same size");
+	  }
+	
 	if(__weight.size())
 	  lookup = KTfwd::fwdpp_internal::gsl_ran_discrete_t_ptr(gsl_ran_discrete_preproc(__weight.size(),&__weight[0]));
       }
       /*!
 	Returns a position from a region that is chosen based on region weights.
+
+	\precondition If recrate == 0, then beg and end cannot be empty.  It is up to the calling environment to make sure
+	that this cannot happen.  This is checked in debug mode via the assert macro.
       */
       template<typename gamete_t,
 	       typename mcont_t>
@@ -176,8 +207,10 @@ namespace KTfwd
 				    const gamete_t &,
 				    const mcont_t &) const
       {
+	assert( !(recrate==0. && (beg.empty()||end.empty())) );
 	auto nbreaks = gsl_ran_poisson(r,recrate);
 	if(!nbreaks) return {};
+
 	result_type rv;
 	for( unsigned i=0;i<nbreaks;++i )
 	  {